% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks-data.R
\docType{data}
\name{parks}
\alias{parks}
\title{Theme park data on WaitTimes}
\format{
An object of class \code{"data.frame" }
}
\source{
\href{queue-times.com/en-US}{Queue-Times.com}
}
\usage{
data(parks)
}
\description{
Data from \href{queue-times.com/en-US}{Queue-Times.com} from all the theme
parks available. Data includes name and location data.
}
\examples{
data(parks)
}
\keyword{datasets}
