% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitTimes.R
\name{waitTimes}
\alias{waitTimes}
\title{Get wait times}
\usage{
waitTimes(parkid, dates)
}
\arguments{
\item{parkid}{The park id for the park you'd like to see}

\item{dates}{A two element character vector date range}
}
\value{
A dataframe
}
\description{
This function will pull wait times for all rides from a given park. The data
comes from . See all the available theme parks using the \code{data(parks)}.
Will need internet to run.
}
\examples{
waitTimes("66", "01-02-2023")

}
